Method to summarize OCMB information in processmrw
Created a function to print out the parsed results of the memory
subsystem for easy digestion.  Note that there is already a dimm
based function but it doesn't seem to work anymore.  It was left
in place for future improvement/reference.

I also tweaked the xml to remove a deprecated value from what we
push out to common-mrw-xml just to clean up the mrw xml a little.

Change-Id: I41e07454091a763e9e092f22507de7a5bb8ab000
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/188225
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: MIKE VAGELINE <mpvageli@us.ibm.com>
Reviewed-by: Shawn Hanson <shanson2@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>