Move some PDR exchange code to extended PLDM module
This commit moves the getRemotePdrRepository() function and its
helper functions from the PLDM base code to the PLDM extended
code.  Since this function is run after targeting is up and
the isteps have started, it does not need to be part of the
PLDM base code.  It can be pulled in, as needed, from the
extended PLDM module like other Hostboot Extended (aka HBI) code.

This change is being done to reduce the size of the Hostboot Base
(aka HBB) image to make room for a future commit that will be
adding new V3 security algorithms.

JIRA: PFHB-677
Change-Id: Ia34dd932f494781d028d08913f37f6393c6a3bee
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/188309
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>