Enable Full V3 Verification In Hostboot Base Code
This code updates the Secure PNOR path to (1) process the HB_HLL
partition and (2) use the sha3() hash values in the HB_HLL
partition to verify other PNOR sections as they are loaded.

To accomplish this, the full V3 verification path was enabled
in the Hostboot Base code (aka HBB).  Any testcases that relied
on V1 verification had to be disabled.

JIRA: PFHB-680
CMVC-Coreq: 1189342
Change-Id: I9fd3f4cdc9a4e4e197533a84f3da9ac93069d31a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/188739
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>