Commonize creating Hash Page Tables for PNOR sections
This commit commonizes the creation of hash page tables for PNOR
partitions.  This will allow the same raw data of a PNOR section
to create the same hash page table, even if it gets rebuilt
multiple times.

This also updates the Secure PNOR HBB->HBI check.  This check will
be skipped when the system is in V3 signing mode because the
HB_HLL checks will provide the same coverage.  For V1 systems,
the check has moved from using the HBB's SW Signature from the V1
secure header to using the the sha3() has of the payload hash
from the same header.

Change-Id: Iec1e37ecb7d5539bc65fe22973ee296a06401a3e
JIRA: PFHB-680
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/188931
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>