Skip V3 HB_HLL verification of some partitions on BMC builds
This commit is a workaround for a known build issue where some
system-specific partitions (aka sections) are not getting properly
set up in Hostboot's HB_HLL partition.  This leads to fails when
the HB_HLL is used for V3 verification.

This commit puts in a temporary workaround for the affected partitions
for BMC builds.

STGD: 698037
Change-Id: I927780f86e0989a9570a2f30b99d0dffd8ff8efe
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/189097
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Matt Strom <matt.strom@ibm.com>
Reviewed-by: Daniel Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>