Add UNIT_SPARED to enum for ATTR_RECONFIGURE_LOOP
There was an issue where error logs which had their severity reduced due
to spare core actions but the istep log associated with it would not
have its severity changed. This commit adds a new enum option to
indicate that a reconfig loop should occur for these actions which will
trigger the istep log to have its severity reduced when spare actions
are taken.

Change-Id: I26c9f1fa86a348e90305eeb6593f2d4bf32f3a2c
STGD:698445
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hw/ekb/+/189921
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joseph McGill <jmcgill@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/189953
Reviewed-by: Daniel Crowell <dcrowell@us.ibm.com>
Reviewed-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Daniel Crowell <dcrowell@us.ibm.com>