Create unique hb_hll files for FSP systems
On FSP systems there are multiple unique targeting binaries that
feed into the HBD pnor partition processing. These targeting
binaries are unique for systems (Denali versus McKinley) and
the different nodes in the system.

The unique HBD per system and nodes means that we also need a
unique HB_HLL with the hash for this unique HBD.

This change lays the groundwork to run a new script to generate
these files but does not actually do this yet.  As a simple
placeholder the new files will be created as copies of the
single partially-correct hb_hll that we currently have.
This will enable some of the FSP work to start in parallel.

In order to make things more clear in the build tree, another
change will use a prefix similar to the original input file
rather than a constanctly incrementing number.

Change-Id: Ieac8c5dfe40ff2f4cd48c1f4b822a4923a9fd7e6
STGD: 702137
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/190460
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>