Fix FSP builds for SBE, HCODE, and HB_HLL creation
This commit fixes a previous commit that was building the SBE and
HCODE PNOR sections/partitions before they were created in a
backing build. (It worked locally in a sandbox.)

It also makes sure that the HB_HLL section is an input into
genPnorImages.pl in multiple places.  This allows HB_HLL to get
built if just the first "build" pass is made, but then also properly
get rebuilt if "intall_all" pass is made.

STGD: 700776
Change-Id: I303d13faac3e14d1a0a61825cdc57fe72da993d1
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/189709
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>