PRD: Remove chip mark cleanup at end of TPS analysis
The chip mark cleanup call at the end of TPS is a hold over from before
row repair support when we may have written a chip mark during TPS but
not subsequently triggered a VCM procedure. As is, it may cause a
timing issue if we get an MPE sometime after TPS is added to the
targeted diagnostics queue but before TPS has finished as we'll take
cleanup actions on that new, unverified chip mark. We should be able
to remove the cleanup call as all chip marks added during TPS should
trigger a VCM event now.

Change-Id: I81b202d2083f5c16ec6addc22a93ca8d49b77c65
STGD: 704372
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/prd/+/190513
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zane Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/190595