Add HBD decompressor to fix attribute updates
During MPIPL/HBRT updates, the PNOR/LID binary has to be reconciled with
the reserved memory copy. In doing that, a temp buffer is created to
hold all the answers but the problem was it didn't account for the HBD
binary not having filled in zero init sections. As a result, any
sections beyond those were misaligned during the compare and would be
susceptible to corruption by the update algorithms.

This commit fixes this issue by adding a decompression algorithm which
will properly place all the data where it ought to have been if the
zero-init sections were taking up their full space.

Change-Id: Ic3df08991aac36d6275235af70470feccd7e16ab
STGD: 701842
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/192223
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel Crowell <dcrowell@us.ibm.com>