Clear Memory Regions Before TCE Allocation
On FSP systems - not BMC systems - Hostboot sets up the TCE regions
for the FSP to load the PAYLOAD (ie, PHYP) and HDAT sections. This
commit clears these memory regions by "memset"-ting them to zero
before settings up the TCEs.

This commit also clears an additional region which is used to store
the PHYP HLL lid.  Clearing this region ensures that the PHYP HLL
will be re-loaded on all normal IPLs and MPIPLs.

STGD: 708514
Change-Id: Ie64a26e9e3515ba161fe6d6aed24c61b63a0e070
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/191988
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Veena Ganti <veena@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel Crowell <dcrowell@us.ibm.com>
Reviewed-by: Matt Strom <matt.strom@ibm.com>