Clear any non-zero data between sections in the SBE Image
This commit looks for and clears any non-zero data in the gaps
between the XIP sections that make up the SBE Image.

While this data was not being used as part of the SBE Image, the
non-zero data was changing the Data CRC that Hostboot calculates
across the entire customized image.  This change in the Data CRC
was causing unnecessary SBE Updates.  Therefore, to reduce spurious
code updates and improve IPL performance, this non-zero data is now
cleared (ie, set to zero).

STGD: 694024
JIRA: PFHB-973
Change-Id: If18094e6d2f2adb081f00476d589f9cf58a9b873
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/192475
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Veena Ganti <veena@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Victor Umoren-Udo <umorenudovictor@ibm.com>
Reviewed-by: Daniel Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>