Ensure that PHYP HLL lid gets loaded on each IPL
This commit ensures that at the start of every IPL - either
a normal IPL or a MPIPL - that the HLL gets loaded from the
FSP or BMC.  This ensures that on an IPL after a code update
at runtime the HLL has valid information for any lids that
might have been updated during the code update.

This commit also adds some additional traces for the HLL path.

STGD: 725875
Change-Id: I4bff42a72310f0d0cd05054813fe7c6dbd3dac9e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/195406
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>