New method to determine PNOR Version for SBE Update
This commit updates how the "PNOR Version" is created for
SBE Updates.

Previously, the "PNOR Version" came from the uncustomized SBE
build image, but it was unique for each build. This was
preventing us from reducing spurious code updates by not
updating the SBEs if the true content had not changed.

Now, the "PNOR Version" is calculated by putting the uncustomized
SBE image into memory and running the maskUnmaskMetaData()
function against it.  Then a sha3() hash is calculated across the
image.

JIRA: PFHB-1048
STGD: 739286
Change-Id: Ib01ecd00e9cdc0aabda68aa082c56c0408ab74fb
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/195820
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>