HB: Move IPL time tracking code to extended image
This commit moves all of the functions that use the global
g_ipl_stats std::array to keep track of istep (and substep)
IPL times from libinitservice.so module in the Hostboot
Base Image (aka HBB) to the libistepdisp.so module in the
Hostboot Extended Image.

This removes ~100KB from the very size-limited Hostboot Base
Image without affecting the results.

This also removes two leftover comments from a previous commit.

Change-Id: Id86ab60a9ae1f4378e39d33377b1c0250af32df2
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/195724
Reviewed-by: Daniel Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Veena Ganti <veena@us.ibm.com>