PRD: add level 2 callout to ME=0 checkstop
After discussion of a field failure it was decided to add a level 2
callout at medium priority so that additional investigation is done to
determine root cause before replacing a part. The processor FRU callout
is still required by manufacturing. However, it has been moved to low
priority because it is least likely that the processor reporting this
attention actually is at fault.

Change-Id: I278d757b3efd0ece6cb07a7445534cd4460a2a68
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/prd/+/196125
Reviewed-by: Zane Shelley <zshelle@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Caleb Palmer <cnpalmer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/196298