HB: Update FAPINAME for OCC
getFapiName() generates an incorrect FAPI_NAME default
values (NA) for attrs in the xmls for OCCs, since they are of
CLASS_UNIT. The error is that FAPI_NAMEs are used to get the
position of a target for attr overrides. This causes an error
'getAttrTankTargetPosData: Error decoding class 0x4:'
This bug causes some attribute overrides to crash Hostboot.

STGD:743069
Change-Id: I1db5bc35fa56a656b208cf6822847bb306bab311
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/hostboot/+/196164
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>